builtin/cat-file.c: support `--[no-]dedup`
Nearly since its inception, `git cat-file --batch-all-objects`'s
behavior is to de-duplicate object IDs before emitting them to the
caller. This behavior dates back to 3115ee45c8 (cat-file: sort and
de-dup output of --batch-all-objects, 2015-06-22), where the rationale
is given as:

    [...] printing duplicates is just a hassle for the user, who must
    then de-dup themselves (or risk a wrong answer if they are doing
    something like counting objects with a particular property).

And indeed, most of the time it is a hassle for the user to have to
de-duplicate the results (or structure whatever is downstream of `git
cat-file` to be able to handle duplicate entries).

But in certain instances, particularly in large repositories, the
benefit outweighs the significant memory cost of holding onto all of the
object IDs either in an `oidset` or `oid_array`.

For example, on linux.git, peak memory usage can exceed 580 MiB, the
vast majority of which is used to store the OIDs of already-emitted
objects. (According to Massif, 99.26% of allocations come from the call
to `oidset_insert()`).

But when the caller can either de-duplicate themselves, or does not care
about duplicate entries, the memory usage drops to just over 90 KiB,
where predictably, most allocations come from inflating objects.

As a nice side-benefit, we can avoid a significant amount of time by not
having to sort and/or de-duplicate the object IDs ahead of time:

    $ ARGS="git.compile cat-file --batch-check='%(objectname)' --batch-all-objects" \
      hyperfine \
        -n dedup "$ARGS" \
        -n no-dedup "$ARGS --no-dedup"
    Benchmark 1: dedup
      Time (mean ± σ):      1.203 s ±  0.021 s    [User: 1.127 s, System: 0.075 s]
      Range (min … max):    1.174 s …  1.240 s    10 runs

    Benchmark 2: no-dedup
      Time (mean ± σ):     658.1 ms ±  17.6 ms    [User: 641.9 ms, System: 16.0 ms]
      Range (min … max):   637.1 ms … 692.1 ms    10 runs

    Summary
      'no-dedup' ran
        1.83 ± 0.06 times faster than 'dedup'

Signed-off-by: Taylor Blau <me@ttaylorr.com>