t1006: extract common setup for `--batch-all-objects` tests
The first test for `--batch-all-objects` sets up the repository in
addition to testing `git cat-file --batch-all-objects`. The tests below
it involving `--batch-all-objects` all also rely on that setup in order
to function.

Split out the setup from the actual test. This has a couple of benefits:

  - it makes it easier to run some `--batch-all-objects` test, but not
    others
  - it isolates failures in the setup from (legitimate) failures due to
    bugs in `--batch-all-objects`

While we're here, drop an instance of `git` on the left-hand side of a
pipe.

Signed-off-by: Taylor Blau <me@ttaylorr.com>